+++ Client/scripts/globals.ts
@@ -1,41 +1,47 @@
 document.onkeydown = (event: KeyboardEvent) => {
-    const key = event.key.toLowerCase();
-    if (key === "f5") {
-        event.returnValue = false;
-        return false;
-    } else if (key === "r" && (event.ctrlKey || event.metaKey)) {
-        event.returnValue = false;
-        return false;
+    if (event instanceof KeyboardEvent){
+        const key = event.key.toLowerCase();
+        if (key === "f5") {
+            event.returnValue = false;
+            return false;
+        } else if (key === "r" && (event.ctrlKey || event.metaKey)) {
+            event.returnValue = false;
+            return false;
+        }
     }
 };
 
-function SetPlayerUID(uid: string) {
+async function SetPlayerUID(uid: string) {
     localStorage.setItem("PlayerUID", uid);
+    return;
 }
 function GetPlayerUID() {
     return localStorage.getItem("PlayerUID");
 }
 
-function ClearStorage() {
+async function ClearStorage() {
     localStorage.clear();
-    localStorage.clear();
+    sessionStorage.clear();
+    return;
 }
 
-function CopyToClipboard(value: string) {
+async function CopyToClipboard(value: string) {
     if ("clipboard" in navigator) {
         navigator.clipboard.writeText(value);
     }
+    return;
 }
 
 function ForceHome() {
     location.href = location.origin;
 }
 
-function FocusElement(selector: string) {
+async function FocusElement(selector: string) {
     const el: HTMLElement = document.body.querySelector(selector);
     if (el) {
         el.focus();
     }
+    return;
 }
 
 function Debug(thing: any) {
@@ -43,7 +49,7 @@ function Debug(thing: any) {
 }
 
 async function GetGridSize(url: string, customSize:number) {
-    let grid = [0, 0];
+    let grid = [0, 0, 0, 0];
     if (url.length) {
         grid = await new Promise((resolve) => {
             const tempImg = document.createElement("img");
@@ -54,11 +60,11 @@ async function GetGridSize(url: string, customSize:number) {
                 const width = Math.floor(bounds.width / customSize);
                 const height = Math.floor(bounds.height / customSize);
                 tempImg.remove();
-                resolve([width, height]);
+                resolve([width, height, bounds.width, bounds.height]);
             });
             tempImg.addEventListener("error", () => {
                 tempImg.remove();
-                resolve([0, 0]);
+                resolve([0, 0, 0, 0]);
             });
             document.body.appendChild(tempImg);
         });
@@ -66,14 +72,15 @@ async function GetGridSize(url: string, customSize:number) {
     return grid;
 }
 
-function ClearHighlightedCells() {
+async function ClearHighlightedCells() {
     const cells = Array.from(document.body.querySelectorAll("td.highlight"));
     for (let i = 0; i < cells.length; i++) {
         cells[i].className = "";
     }
+    return;
 }
 
-function PlaySound(name:string){
+async function PlaySound(name:string){
     switch(name){
         case "alert.wav":
             if (!localStorage.getItem("alertDisabled")){
@@ -105,6 +112,7 @@ function PlaySound(name:string){
             audio.play();
             break;
     }
+    return;
 }
 
 function Ping(x:number, y:number){
@@ -136,9 +144,10 @@ function Ping(x:number, y:number){
     }, 2000);
 }
 
+let tabletop:HTMLElement;
 function DragTabletop(){
-    const tabletop:HTMLElement = document.body.querySelector(".js-tabletop");
-    if (tabletop){
+    if (!tabletop){
+        tabletop = document.body.querySelector(".js-tabletop");
         let pos = { top: 0, left: 0, x: 0, y: 0 };
         let movingTabletop = false;
         tabletop.addEventListener("mousedown", (e:MouseEvent)=>{