+++ Shared/Models/Tabletop.cs
@@ -13,12 +13,13 @@ namespace FreeTabletop.Shared.Models
         public List<NPC> NPCs = new List<NPC>();
         public List<Entity> CombatOrder = new List<Entity>();
         public string UID { get; set; }
-        public string GridType = "1";
+        public string GridType = null;
         public string Image = null;
         public int[] Grid { get; set; }
         public List<Message> Messages = new List<Message>();
-        public string GameMasterUID {get;set;}
-        public int CellSize;
+        public string GameMasterUID { get; set; }
+        public int CellSize { get; set; }
+        public int[] Size { get; set; }
 
         public void MoveLocalEntitiy(string uid, int[] newPosition)
         {